LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/reading-scroll-forces-anchoring-expected.txt

CONSOLE MESSAGE: Error: assert_equals: expected 250 but got 150
abc
def

Harness Error (FAIL), message = Error: assert_equals: expected 250 but got 150

TIMEOUT Reading scroll position forces scroll anchoring adjustment. Test timed out

Harness Error (FAIL), message = Error: assert_equals: expected 250 but got 150

TIMEOUT Reading scroll position forces scroll anchoring adjustment. Test timed out
